Upgrades the repo to tf 2.4 and change the data loading setup to use a dataset generated directly from the input data file rather than creating an intermediate tfrecord file
We eliminate the intermediate tfrecord file by creating a from_generator Dataset from data read in from the tsv input file rather than first creating a tfrecord file and then creating a Dataset from that. To get a from_generator Dataset to properly output sparse tensors as the model requires, we upgrade the code to tf 2.x and make adjustments to the model accordingly. In particular, we have to change the LazyAdamOptimizer to one from a different library as the tf1 contrib library is not longer supported. Ideally this should have been 2 commits, but after starting on moving the Dataset to from_generator, I realized tf2 was needed.